tests

+++ reseptisovellus/src/Components/Pages/RecipeDetails.js
@@ -4,9 +4,6 @@ import "./RecipeDetails.css";
 import { StarRating } from "../StarRating";
 import { useAuth } from "../Authentication";
 import { NavLink } from "react-router-dom";
-
-
-
 import {
   EmailShareButton,
   FacebookShareButton,
@@ -301,32 +298,38 @@ const RecipeDetails = () => {
   };
 
   // function for admins to delete recipe
-  const deleteRecipe = async () => {
-    try {
-      const response = await fetch(
-        // `https://recipeappapi.azurewebsites.net/api/recipe/${id}`,
-        `https://localhost:7005/api/recipe/${id}`,
-        {
-          method: "DELETE",
-          headers: {
-            Authorization: `Bearer ${authUser.token}`,
-          },
-        }
-      );
-
-      if (response.ok) {
-        console.log("Recipe deleted successfully.");
-        alert("Recipe deleted successfully.");
-        //window.location.reload();
-        navigate("/recipes");
-      } else {
-        console.error("Error deleting recipe:", response);
-      }
-      
-    } catch (error) {
-      console.error("Error deleting recipe:", error);
+const deleteRecipe = async () => {
+  try {
+    // Kysytään käyttäjältä varmistus poistosta
+    const confirmDelete = window.confirm("Are you sure you want to delete this recipe?");
+    if (!confirmDelete) {
+      return; // Poistetaan toiminto, jos käyttäjä ei vahvista poistoa
     }
-  };
+
+    const response = await fetch(
+      // `https://recipeappapi.azurewebsites.net/api/recipe/${id}`,
+      `https://localhost:7005/api/recipe/${id}`,
+      {
+        method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${authUser.token}`,
+        },
+      }
+    );
+
+    if (response.ok) {
+      console.log("Recipe deleted successfully.");
+      alert("Recipe deleted successfully.");
+      //window.location.reload();
+      navigate("/recipes");
+    } else {
+      console.error("Error deleting recipe:", response);
+    }
+    
+  } catch (error) {
+    console.error("Error deleting recipe:", error);
+  }
+};
 
   const currentPage = `https://recipeappgl.azurewebsites.net/`;
 
@@ -394,7 +397,7 @@ const RecipeDetails = () => {
           </div>
 
           <div className="recipe-detail-image">
-            {recipe.images.length > 0 ? (
+            {recipe.images && recipe.images.length > 0 ? (
               <RecipeSlider images={recipe.images} />
             ) : (
               <img src="/default_pic.jpg" alt="default picture" />
@@ -403,7 +406,7 @@ const RecipeDetails = () => {
 
           <p>
             Keywords:{" "}
-            {recipe.keywords.map((keyword) => keyword.word).join(", ")}
+            {recipe.keywords && recipe.keywords.map((keyword) => keyword.word).join(", ")}
           </p>
 
           <div className="recipe-detail-actions">
